refactor: 지원하지 않는 console UI 삭제
