+++ src/main/java/roomescape/controller/AdminViewController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.web;
+package roomescape.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
